requirements.txt
torch==1.8.1
opencv_python==4.5.1.48
pandas==1.2.4
pdf2image==1.14.0
pytesseract==0.3.7
streamlit==0.80.0
pytorch_lightning==1.2.10
numpy==1.20.2
torchvision==0.9.1
albumentations==0.5.2
scikit_image==0.18.1
Pillow==8.2.0
skimage==0.0
